Тестирование файлов

+++ app/Http/Controllers/Admin/AdminArticlesFilesController.php
@@ -30,8 +30,40 @@ class AdminArticlesFilesController extends Controller{
 	public function update(Request $request, Article $article){
 	}
 
+	public function upload(Article $article, Request $request){
+		$this->saveFile($article, $request->file('file'));
+	}
+
+	public function saveFile(Article $article, UploadedFile $file){
+		$name = basename($file->getClientOriginalName());
+		$extension = $file->getClientOriginalExtension();
+		$newName = translit($name); // точка не исчезает
+		$path = $article->filesFolder.DIRECTORY_SEPARATOR;
+		// dd($newName);
+
+		// проверка на уникальность
+		if (ArticleFile::where('path', $path.$newName)->count()){
+			return ArticleFile::where('path', $path.$newName)->get();
+		}
+
+		// сохраняем файл
+		$file->move(public_path().$path, $newName);
+
+		// создаём запись в бд
+		return ArticleFile::create([
+			'article_id'    => $article->id,
+			'path'          => $path.$newName,
+			'title'         => '',
+			'alt'           => '',
+			'extension'     => $extension,
+			'original_name' => $name,
+		]);
+	}
+
 	public function destroy(Article $article, ArticleFile $file){
-		$file->deleteFile();
+		if (file_exists(public_path().$file->path)){
+			$file->deleteFile();
+		}
 		$file->delete();
 
 		return back();
